Organize org.jboss.as.clustering.infinispan classes into packages.
