+++ clustering/infinispan/extension/src/main/java/org/jboss/as/clustering/infinispan/manager/DefaultCacheContainerAdmin.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.jboss.as.clustering.infinispan;
+package org.jboss.as.clustering.infinispan.manager;
 
 import java.util.EnumSet;
