+++ clustering/infinispan/extension/src/main/java/org/jboss/as/clustering/infinispan/subsystem/TransactionServiceConfigurator.java
@@ -36,10 +36,10 @@ import org.infinispan.configuration.cache.TransactionConfiguration;
 import org.infinispan.configuration.cache.TransactionConfigurationBuilder;
 import org.infinispan.transaction.LockingMode;
 import org.infinispan.transaction.tm.EmbeddedTransactionManager;
-import org.jboss.as.clustering.infinispan.TransactionManagerProvider;
-import org.jboss.as.clustering.infinispan.TransactionSynchronizationRegistryProvider;
 import org.jboss.as.clustering.infinispan.jakarta.TransactionManagerAdapter;
 import org.jboss.as.clustering.infinispan.jakarta.TransactionSynchronizationRegistryAdapter;
+import org.jboss.as.clustering.infinispan.tx.TransactionManagerProvider;
+import org.jboss.as.clustering.infinispan.tx.TransactionSynchronizationRegistryProvider;
 import org.jboss.as.controller.OperationContext;
 import org.jboss.as.controller.OperationFailedException;
 import org.jboss.as.controller.PathAddress;
