+++ clustering/infinispan/extension/src/main/java/org/jboss/as/clustering/infinispan/tx/InfinispanXAResourceRecovery.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.jboss.as.clustering.infinispan;
+package org.jboss.as.clustering.infinispan.tx;
 
 import javax.transaction.xa.XAResource;
